Boot Config: Move into submodule
Consolidate the boot configuration back-end tools into a submodule.
