+++ kano_settings/system/boot_config/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+
+# __init__.py
+#
+# Copyright (C) 2014 Kano Computing Ltd.
+# License: http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
+#
+
+__author__ = 'Kano Computing Ltd.'
+__email__ = 'dev@kano.me'
