+++ kano_settings/system/boot_config/boot_config_parser.py
@@ -1,7 +1,7 @@
 import re
 
-from kano_settings.boot_config_filter import Filter
-from kano_settings.boot_config_line import BootConfigLine
+from kano_settings.system.boot_config.boot_config_filter import Filter
+from kano_settings.system.boot_config.boot_config_line import BootConfigLine
 
 
 class BootConfigParser(object):
